refactor: extract features loading to common helpers
